+++ src/main/java/uk/gov/di/ipv/cri/common/library/domain/personidentity/PersonIdentityDetailed.java
@@ -19,20 +19,60 @@ public class PersonIdentityDetailed {
     @JsonProperty("drivingPermit")
     private final List<DrivingPermit> drivingPermits;
 
-    public PersonIdentityDetailed(
-            List<Name> names, List<BirthDate> birthDates, List<Address> addresses) {
-        this(names, birthDates, addresses, null);
+    @JsonProperty("passport")
+    private final List<Passport> passports;
+
+    /**
+     * @param names list of names
+     * @param birthDates list of birthDates
+     * @deprecated Avoid use outside CRI-Lib, Restricted section of audits should be made generic in
+     *     future and provided by the CRI To avoid the current pattern of adding CRI specific fields
+     *     to this class
+     * @see uk.gov.di.ipv.cri.common.library.service.PersonIdentityDetailedFactory
+     *     createPersonIdentityDetailedWith methods
+     */
+    @Deprecated(since = "1.5.0")
+    public PersonIdentityDetailed(List<Name> names, List<BirthDate> birthDates) {
+        this(names, birthDates, null, null, null);
     }
 
+    /**
+     * @param names list of names
+     * @param birthDates list of birthDates
+     * @param addresses list of addresses, null list if not used
+     * @deprecated Avoid use outside CRI-Lib.
+     * @see uk.gov.di.ipv.cri.common.library.service.PersonIdentityDetailedFactory
+     *     createPersonIdentityDetailedWith methods
+     */
+    @Deprecated(since = "1.5.0")
+    public PersonIdentityDetailed(
+            List<Name> names, List<BirthDate> birthDates, List<Address> addresses) {
+        this(names, birthDates, addresses, null, null);
+    }
+
+    /**
+     * @param names list of names
+     * @param birthDates list of birthDates
+     * @param addresses null list if not used
+     * @param drivingPermits null list if not used
+     * @param passports null list if not used
+     * @deprecated Do not use outside CRI-Lib, CRI's should use PersonIdentityDetailedFactory with
+     *     the appropriate createPersonIdentityDetailedWith methods.
+     * @see uk.gov.di.ipv.cri.common.library.service.PersonIdentityDetailedFactory
+     *     createPersonIdentityDetailedWith
+     */
+    @Deprecated(since = "1.5.0")
     public PersonIdentityDetailed(
             List<Name> names,
             List<BirthDate> birthDates,
             List<Address> addresses,
-            List<DrivingPermit> drivingPermits) {
+            List<DrivingPermit> drivingPermits,
+            List<Passport> passports) {
         this.names = names;
         this.birthDates = birthDates;
         this.addresses = addresses;
         this.drivingPermits = drivingPermits;
+        this.passports = passports;
     }
 
     public List<Name> getNames() {
@@ -49,5 +89,9 @@ public class PersonIdentityDetailed {
 
     public List<DrivingPermit> getDrivingPermits() {
         return drivingPermits;
+    }
+
+    public List<Passport> getPassports() {
+        return passports;
     }
 }